Adding RetainedFileCountLimit
The default Seilog RollingFile implementation has the ability to set the RetainedFileCountLimit. This deletes any files over the set limit when the lof file rolls over. 

The default implementation has a default of 31 days (as it only rolls over on a day basis, this implementatiion keeps the existing behaviour of unlimited log files by keeping the default set as null.