Adding RetainedFileCountLimit (#22)
The default Serilog RollingFile implementation has the ability to set the RetainedFileCountLimit. This deletes any files over the set limit when the log file rolls over. 

The default implementation has a default of 31 days (as it only rolls over on a day basis, this implementation keeps the existing behaviour of unlimited log files by keeping the default set as null.