ITextFormatter support for both sinks
Allows for the overloaded option to pass an ITextFormatter from the Serilog framework. This enhances the options of the Alternative Rolling File Sink as it can now output in JSON, Raw, or Display formats. If there are new formats added to the Serilog framework, they would be supported as well. 

This is functionality that exists in the Serilog Rolling File sink and should be extended to the Alternative Rolling File sink.