Roll over when the date changes.
At present, the files get rolled over by size. When the new day starts, the log files written still have the previous date string in the file name until a new file is created on that day. This change will make sure that a new log files are created when the date changes.